refactor multi_peer_conductor hpp => cpp
