+++ Plugins/NativeServerPlugin/StreamingNativeServerPlugin.vcxproj
@@ -189,6 +189,7 @@
     <ClCompile Include="src\directx_buffer_capturer.cpp" />
     <ClCompile Include="src\directx_peer_conductor.cpp" />
     <ClCompile Include="src\input_data_channel_observer.cpp" />
+    <ClCompile Include="src\multi_peer_conductor.cpp" />
     <ClCompile Include="src\peer_view.cpp" />
     <ClCompile Include="src\render_service.cpp" />
     <ClCompile Include="src\service_base.cpp" />
@@ -200,7 +201,7 @@
     <ClInclude Include="inc\directx_buffer_capturer.h" />
     <ClInclude Include="inc\directx_peer_conductor.h" />
     <ClInclude Include="inc\input_data_channel_observer.h" />
-    <ClInclude Include="inc\multi_peer_conductor.hpp" />
+    <ClInclude Include="inc\multi_peer_conductor.h" />
     <ClInclude Include="inc\peer_conductor.h" />
     <ClInclude Include="inc\peer_view.h" />
     <ClInclude Include="inc\plugindefs.h" />
